Implementation Crafting
Items can be combined in Slot 13 and 14 to create a new Item.
Items are moved back into the inventory on failed crafting and removed on successful crafting.
If the crafted Item remains in Slot 15 on Inventory close, it is automatically moved into the inventory.

ToDo
Edge Case: Too many Items