Completed Goal System
By adding the Control Goal Script to an Object
Objects can now, depending on the Setting of the Script,
Add a Goal to the Goal List, or Complete a Goal.
This Action can be triggered following successful Interaction or Dialogue. (It can be expanded to include Locked Interaction maybe)

The Goal List now contains GoalObjects, which remember their position once added to the List and can be either completed or uncompleted.

Using 2 Buttons, the GoalScroll Script allows to scroll up or down the Goal List. The Last Position of the List is stored and remembered by the GoalList Script (has to be transfered to DataManager)

Goals are Sprites.